+++ iblrig/base_choice_world.py
@@ -1,6 +1,7 @@
 """Extends the base_tasks modules by providing task logic around the Choice World protocol."""
 
 import abc
+import enum
 import logging
 import math
 import random
@@ -9,7 +10,7 @@ import time
 from pathlib import Path
 from re import split as re_split
 from string import ascii_letters
-from typing import Annotated, Any
+from typing import Annotated, Any, final
 
 import numpy as np
 import pandas as pd
@@ -29,6 +30,7 @@ log = logging.getLogger(__name__)
 
 NTRIALS_INIT = 2000
 NBLOCKS_INIT = 100
+
 
 # TODO: task parameters should be verified through a pydantic model
 #
@@ -168,37 +170,135 @@ class ChoiceWorldSession(
             self.start_mixin_bonsai_visual_stimulus()
             self.bpod.register_softcodes(self.softcode_dictionary())
 
-    def _run(self):
-        """Run the task with the actual state machine."""
+    @final
+    def _wait_for_camera_and_initial_delay(self) -> None:
+        """Wait for the camera to start recording and manage the initial delay.
+
+        This method implements a temporary state machine to coordinate the process of waiting for the camera recording
+        to commence and to handle any specified initial delay. It should be called just prior to the start of the task.
+        The states defined here were previously part of the task's main state machine (see `get_state_machine_trial()`).
+        """
+        initial_delay = self.task_params.get('SESSION_DELAY_START', 0)
+
+        # temporary IntEnum for storing softcodes
+        # SOFTCODE.TRIGGER_CAMERA is being reused; we add three more unique values
+        class TemporarySoftcodes(enum.IntEnum):
+            START_CAMERA_RECORDING = SOFTCODE.TRIGGER_CAMERA.value
+            WAIT_FOR_CAMERA_TRIGGER = enum.auto()
+            CAMERA_TRIGGER_RECEIVED = enum.auto()
+            STARTING_INITIAL_DELAY = enum.auto()
+
+        # store the original softcode handler
+        original_softcode_handler = self.bpod.softcode_handler_function
+
+        # define temporary softcode handler
+        def temporary_softcode_handler(softcode: int):
+            match softcode:
+                case TemporarySoftcodes.START_CAMERA_RECORDING:
+                    original_softcode_handler(softcode)  # pass to original handler
+                case TemporarySoftcodes.WAIT_FOR_CAMERA_TRIGGER:
+                    log.info('Waiting to receive first camera trigger ...')
+                case TemporarySoftcodes.CAMERA_TRIGGER_RECEIVED:
+                    log.info('Camera trigger received')
+                case TemporarySoftcodes.STARTING_INITIAL_DELAY:
+                    if initial_delay > 0:
+                        log.info(f'Waiting for {initial_delay} s')
+                    else:
+                        log.info('No initial delay defined')
+
+        # overwrite softcode handler
+        self.bpod.softcode_handler_function = temporary_softcode_handler
+
+        # define and run state machine
+        sma = StateMachine(self.bpod)
+        sma.add_state(
+            state_name='start_camera_workflow',
+            output_actions=[('SoftCode', TemporarySoftcodes.START_CAMERA_RECORDING)],
+            state_change_conditions={'Tup': 'wait_for_camera_trigger'},
+        )
+        sma.add_state(
+            state_name='wait_for_camera_trigger',
+            output_actions=[('SoftCode', TemporarySoftcodes.WAIT_FOR_CAMERA_TRIGGER)],
+            state_change_conditions={'Port1In': 'camera_trigger_received'},
+        )
+        sma.add_state(
+            state_name='camera_trigger_received',
+            output_actions=[('SoftCode', TemporarySoftcodes.CAMERA_TRIGGER_RECEIVED)],
+            state_change_conditions={'Tup': 'delay_initiation'},
+        )
+        sma.add_state(
+            state_name='delay_initiation',
+            state_timer=self.task_params.get('SESSION_DELAY_START', 0),
+            output_actions=[('SoftCode', TemporarySoftcodes.STARTING_INITIAL_DELAY)],
+            state_change_conditions={'Tup': 'exit'},
+        )
+        self.bpod.send_state_machine(sma)
+        self.bpod.run_state_machine(sma)  # blocking until state-machine is finished
+        if initial_delay > 0:
+            log.info('Initial delay has passed')
+
+        # restore original softcode handler
+        self.bpod.softcode_handler_function = original_softcode_handler
+
+    def _run(self) -> None:
+        """Execute the task using the defined state machine.
+
+        This method orchestrates the execution of the task by running a state machine for a specified number of trials.
+        """
         time_last_trial_end = time.time()
-        for i in range(self.task_params.NTRIALS):  # Main loop
-            # t_overhead = time.time()
+        for trial_number in range(self.task_params.NTRIALS):  # Main loop
+            # obtain state machine definition
             self.next_trial()
-            log.info(f'Starting trial: {i}')
-            # =============================================================================
-            #     Start state machine definition
-            # =============================================================================
-            sma = self.get_state_machine_trial(i)
+            sma = self.get_state_machine_trial(trial_number)
+
+            # Waiting for camera / initial delay will be handled just prior to the first trial
+            # This is done here to allow for backward compatibility with unadapted tasks
+            if trial_number == 0:
+                # warn if state machine uses deprecated way of waiting for camera / initial delay
+                if (5, SOFTCODE.TRIGGER_CAMERA) in sma.output_matrix[0] and sma.state_names[1] == 'delay_initiation':
+                    log.warning('')
+                    log.warning('**********************************************')
+                    log.warning('ATTENTION: YOUR TASK DEFINITION NEEDS UPDATING')
+                    log.warning('**********************************************')
+                    log.warning('Camera and initial delay should not be handled')
+                    log.warning('within the `get_state_machine_trial()` method.')
+                    log.warning('For further details, please refer to section  ')
+                    log.warning("'Deprecation Notes' in IBLRIG's documentation.")
+                    log.warning('**********************************************')
+                    log.warning('')
+                    log.info('Waiting for 10s so you actually read this message ;-)')
+                    time.sleep(10)
+                else:
+                    self._wait_for_camera_and_initial_delay()
+
+            # send state machine description to Bpod device
             log.debug('Sending state machine to bpod')
-            # Send state machine description to Bpod device
             self.bpod.send_state_machine(sma)
-            # t_overhead = time.time() - t_overhead
-            # The ITI_DELAY_SECS defines the grey screen period within the state machine, where the
-            # Bpod TTL is HIGH. The DEAD_TIME param defines the time between last trial and the next
-            dead_time = self.task_params.get('DEAD_TIME', 0.5)
-            dt = self.task_params.ITI_DELAY_SECS - dead_time - (time.time() - time_last_trial_end)
-            # wait to achieve the desired ITI duration
-            if dt > 0:
-                time.sleep(dt)
-            # Run state machine
+
+            # handle ITI durations
+            if trial_number > 0:
+                # The ITI_DELAY_SECS defines the grey screen period within the state machine, where the
+                # Bpod TTL is HIGH. The DEAD_TIME param defines the time between last trial and the next
+                dead_time = self.task_params.get('DEAD_TIME', 0.5)
+                dt = self.task_params.ITI_DELAY_SECS - dead_time - (time.time() - time_last_trial_end)
+
+                # wait to achieve the desired ITI duration
+                if dt > 0:
+                    log.debug(f'Waiting {dt} s to achieve an ITI duration of {self.task_params.ITI_DELAY_SECS} s')
+                    time.sleep(dt)
+
+            # run state machine
+            log.info('-----------------------')
+            log.info(f'Starting Trial #{trial_number}')
             log.debug('running state machine')
             self.bpod.run_state_machine(sma)  # Locks until state machine 'exit' is reached
             time_last_trial_end = time.time()
+
             # handle pause event
             flag_pause = self.paths.SESSION_FOLDER.joinpath('.pause')
             flag_stop = self.paths.SESSION_FOLDER.joinpath('.stop')
-            if flag_pause.exists() and i < (self.task_params.NTRIALS - 1):
-                log.info(f'Pausing session inbetween trials {i} and {i + 1}')
+            if flag_pause.exists() and trial_number < (self.task_params.NTRIALS - 1):
+                log.info(f'Pausing session inbetween trials {trial_number} and {trial_number + 1}')
                 while flag_pause.exists() and not flag_stop.exists():
                     time.sleep(1)
                 self.trials_table.at[self.trial_num, 'pause_duration'] = time.time() - time_last_trial_end
@@ -211,7 +311,7 @@ class ChoiceWorldSession(
 
             # handle stop event
             if flag_stop.exists():
-                log.info('Stopping session after trial %d', i)
+                log.info('Stopping session after trial %d', trial_number)
                 flag_stop.unlink()
                 break
 
@@ -257,6 +357,7 @@ class ChoiceWorldSession(
                 'bonsai_closed_loop': daction,
                 'bonsai_freeze_stim': daction,
                 'bonsai_show_center': daction,
+                'bonsai_freeze_center': daction,
             }
         )
 
@@ -285,9 +386,9 @@ class ChoiceWorldSession(
             if ~np.isnan(sma.state_timer_matrix[i]):
                 out_state = states_indices[sma.state_timer_matrix[i]]
                 edges.append(f'{letter}{states_letters[out_state]}')
-            for input in sma.input_matrix[i]:
-                if input[0] == 0:
-                    edges.append(f'{letter}{states_letters[states_indices[input[1]]]}')
+            for inputs in sma.input_matrix[i]:
+                if inputs[0] == 0:
+                    edges.append(f'{letter}{states_letters[states_indices[inputs[1]]]}')
         dot.edges(edges)
         if output_file is not None:
             try:
@@ -303,30 +404,13 @@ class ChoiceWorldSession(
         # we define the trial number here for subclasses that may need it
         sma = self._instantiate_state_machine(trial_number=i)
 
-        if i == 0:  # First trial exception start camera
-            session_delay_start = self.task_params.get('SESSION_DELAY_START', 0)
-            log.info('First trial initializing, will move to next trial only if:')
-            log.info('1. camera is detected')
-            log.info(f'2. {session_delay_start} sec have elapsed')
-            sma.add_state(
-                state_name='trial_start',
-                state_timer=0,
-                state_change_conditions={'Port1In': 'delay_initiation'},
-                output_actions=[('SoftCode', SOFTCODE.TRIGGER_CAMERA), ('BNC1', 255)],
-            )  # start camera
-            sma.add_state(
-                state_name='delay_initiation',
-                state_timer=session_delay_start,
-                output_actions=[],
-                state_change_conditions={'Tup': 'reset_rotary_encoder'},
-            )
-        else:
-            sma.add_state(
-                state_name='trial_start',
-                state_timer=0,  # ~100µs hardware irreducible delay
-                state_change_conditions={'Tup': 'reset_rotary_encoder'},
-                output_actions=[self.bpod.actions.stop_sound, ('BNC1', 255)],
-            )  # stop all sounds
+        # Signal trial start and stop all sounds
+        sma.add_state(
+            state_name='trial_start',
+            state_timer=0,  # ~100µs hardware irreducible delay
+            state_change_conditions={'Tup': 'reset_rotary_encoder'},
+            output_actions=[self.bpod.actions.stop_sound, ('BNC1', 255)],
+        )
 
         # Reset the rotary encoder by sending the following opcodes via the modules serial interface
         # - 'Z' (ASCII 90): Set current rotary encoder position to zero
@@ -427,7 +511,7 @@ class ChoiceWorldSession(
         sma.add_state(
             state_name='freeze_reward',
             state_timer=0,
-            output_actions=[self.bpod.actions.bonsai_show_center],
+            output_actions=[self.bpod.actions.bonsai_freeze_center],
             state_change_conditions={'Tup': 'reward'},
         )
         sma.add_state(
@@ -583,10 +667,10 @@ class ChoiceWorldSession(
 
         # log info dict
         log.log(log_level, f'Outcome of Trial #{trial_info.trial_num}:')
-        max_key_length = max(len(key) for key in info_dict)
+        key_format = '- {}: '
+        n_justify = max(len(key) for key in info_dict) + len(key_format.format(''))
         for key, value in info_dict.items():
-            spaces = (max_key_length - len(key)) * ' '
-            log.log(log_level, f'- {key}: {spaces}{str(value)}')
+            log.log(log_level, key_format.format(key).ljust(n_justify) + str(value))
 
     @property
     def iti_reward(self):
@@ -655,24 +739,16 @@ class HabituationChoiceWorldSession(ChoiceWorldSession):
     def get_state_machine_trial(self, i):
         sma = StateMachine(self.bpod)
 
-        if i == 0:  # First trial exception start camera
-            log.info('Waiting for camera pulses...')
-            sma.add_state(
-                state_name='iti',
-                state_timer=3600,
-                state_change_conditions={'Port1In': 'stim_on'},
-                output_actions=[self.bpod.actions.bonsai_hide_stim, ('SoftCode', SOFTCODE.TRIGGER_CAMERA), ('BNC1', 255)],
-            )  # start camera
-        else:
-            # NB: This state actually the inter-trial interval, i.e. the period of grey screen between stim off and stim on.
-            # During this period the Bpod TTL is HIGH and there are no stimuli. The onset of this state is trial end;
-            # the offset of this state is trial start!
-            sma.add_state(
-                state_name='iti',
-                state_timer=1,  # Stim off for 1 sec
-                state_change_conditions={'Tup': 'stim_on'},
-                output_actions=[self.bpod.actions.bonsai_hide_stim, ('BNC1', 255)],
-            )
+        # NB: This state actually the inter-trial interval, i.e. the period of grey screen between stim off and stim on.
+        # During this period the Bpod TTL is HIGH and there are no stimuli. The onset of this state is trial end;
+        # the offset of this state is trial start!
+        sma.add_state(
+            state_name='iti',
+            state_timer=1,  # Stim off for 1 sec
+            state_change_conditions={'Tup': 'stim_on'},
+            output_actions=[self.bpod.actions.bonsai_hide_stim, ('BNC1', 255)],
+        )
+
         # This stim_on state is considered the actual trial start
         sma.add_state(
             state_name='stim_on',