+++ iblrig/hardware_validation.py
@@ -395,7 +395,7 @@ class ValidatorBpod(ValidatorSerial):
 
         # check for (un)expected input events
         for event_name, timestamps in sorted(events.items()):
-            if event_name.endswith('Out') or event_name.endswith('Low') or event_name == 'Tup':
+            if event_name.endswith(('Out', 'Low')) or event_name == 'Tup':
                 continue
             diff_timestamps = np.diff(timestamps)
             n_timestamps = len(timestamps)
@@ -495,6 +495,25 @@ class ValidatorCamera(Validator):
             else:
                 yield Result(Status.WARN, f'Measured TTL rate: {trigger_rate:.1f} Hz (expecting {target_rate} Hz)')
         return True
+
+
+class ValidatorDataFolders(Validator):
+    _name = 'Data Folders'
+
+    def _run(self):
+        if self.iblrig_settings.iblrig_remote_data_path in (None, False):
+            yield Result(Status.SKIP, 'iblrig_remote_data_path has not been set in hardware_settings.yaml - skipping validation')
+            return False
+        elif self.iblrig_settings.iblrig_remote_data_path.exists():
+            yield Result(Status.PASS, f"Remote data path '{self.iblrig_settings.iblrig_remote_data_path}' is accessible")
+            return True
+        else:
+            yield Result(
+                Status.FAIL,
+                f"Cannot access remote data path '{self.iblrig_settings.iblrig_remote_data_path}'",
+                solution='Check network connection and mapping of network drive',
+            )
+            return False
 
 
 class ValidatorAlyx(Validator):
@@ -824,7 +843,7 @@ class ValidatorSound(ValidatorSerial):
 
 
 def get_all_validators() -> list[type[Validator]]:
-    return [cast(type[Validator], x) for x in get_inheritors(Validator) if not isabstract(x)]
+    return [cast('type[Validator]', x) for x in get_inheritors(Validator) if not isabstract(x)]
 
 
 def run_all_validators(
@@ -869,7 +888,7 @@ def run_all_validators_cli():
             print(f'{color}  {symbol}  {result.message}{ANSI.END}')
             if result.solution is not None and len(result.solution) > 0:
                 print(f'{color}     Suggestion: {result.solution}{ANSI.END}')
-        print('')
+        print()
     if fail > 0:
         print(ANSI.RED + ANSI.BOLD + f'{fail} validation{"s" if fail > 1 else ""} failed.')
     if warn > 0: